NPC logic and script changes
