+++ Assets/Scenes/Prefabs/Villager1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2b7d89f8671a2ff44870ac96ac6967f9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
