+++ Assets/Scenes/Prefabs/Villager2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9e8697e22746c004a9c8918cea03ac07
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
