+++ Assets/Scenes/Prefabs/Villager3.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f0ac8362cd33bba4787eeba417b110ee
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
