+++ Assets/Scenes/Scripts/Enemies.cs
@@ -6,7 +6,21 @@ public class Enemies : MonoBehaviour
 {
     [SerializeField] protected int lifeQuantity;
     [SerializeField] protected float velocity;
+    [SerializeField] protected float range;
 
+    protected float distanceFromPlayer1;
+    protected float distanceFromPlayer2;
+
+    protected Transform player1Transform;
+    protected Transform player2Transform;
+    
     protected Animator AnimEnemy;
+
     protected virtual void Die() { }
+
+    protected virtual void FindPlayer()
+    {
+        player1Transform = GameObject.FindGameObjectWithTag("Player1").transform;
+        player2Transform = GameObject.FindGameObjectWithTag("Player2").transform;
+    }
 }
