--- Assets/Scenes/Scripts/dialogInterface.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1c271d7974a59624fa5298731dd88951
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
