--- Assets/Scenes/Scripts/enemyFlyer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5c8235f2025841542a432413f049dd29
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
