+++ Assets/Scenes/Scripts/enemyFollow.cs
@@ -2,20 +2,14 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class enemyFollower : Enemies
+public class enemyFollow : Enemies
 {
-    [SerializeField] protected float range;
-    protected float distanceFromPlayer1;
-    protected float distanceFromPlayer2;
-    protected Transform player1Transform;
-    protected Transform player2Transform;
-
-    protected virtual void FindPlayer()
+    void Start()
     {
-        player1Transform = GameObject.FindGameObjectWithTag("Player1").transform;
-        player2Transform = GameObject.FindGameObjectWithTag("Player2").transform;
+        FindPlayer();
     }
-    protected virtual void FollowPlayer() 
+
+    void Update()
     {
         distanceFromPlayer1 = Vector2.Distance(player1Transform.position, transform.position);
         distanceFromPlayer2 = Vector2.Distance(player2Transform.position, transform.position);
@@ -27,6 +21,27 @@ public class enemyFollower : Enemies
         if (distanceFromPlayer2 < range && distanceFromPlayer2 < distanceFromPlayer1)
         {
             transform.position = Vector2.MoveTowards(this.transform.position, player2Transform.position, velocity * Time.deltaTime);
+        }
+    }
+
+    private void FixedUpdate()
+    {
+        Die();
+    }
+
+    private void OnTriggerEnter2D(Collider2D collision)
+    {
+        if (collision.gameObject.CompareTag("lance"))
+        {
+            lifeQuantity--;
+        }
+    }
+
+    protected override void Die()
+    {
+        if (lifeQuantity == 0)
+        {
+            Destroy(this.gameObject);
         }
     }
 }
