--- Assets/Scenes/Scripts/enemyFollower.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dd062ab58521bf14284ec60cb49fd181
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
