+++ Assets/Scenes/Scripts/meleeplayer.cs
@@ -48,7 +48,15 @@ public class meleeplayer : Heroes
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
-        if (collision.gameObject.CompareTag("EnemyFlyer"))
+        if (collision.gameObject.GetComponent<Enemies>())
+        {
+            lifeQuantity--;
+        }
+    }
+
+    private void OnCollisionEnter2D(Collision2D collision)
+    {
+        if (collision.gameObject.GetComponent<Enemies>())
         {
             lifeQuantity--;
         }
