+++ Assets/Scenes/Scripts/scriptableNPC.cs
@@ -7,23 +7,15 @@ using TMPro;
 [CreateAssetMenu(fileName = "NPC Scriptable", menuName = "NPCMenu")]
 public class scriptableNPC : ScriptableObject
 {
-    [SerializeField] private int hp;
     [SerializeField] private string[] dialogs;
-    [SerializeField] private TextMeshProUGUI dialog;
+    [SerializeField] private int timeMax;
 
-    public int HP { get { return hp; } }
     public string[] Dialogs { get { return dialogs; } }
-    public TextMeshProUGUI Dialog { get { return dialog; } }
+    public int TimeMax { get {  return timeMax; } }
 
-    public void Talk()
+
+    public string  Talk()
     {
-        dialog = GameObject.FindGameObjectWithTag("npcDialog").GetComponent<TextMeshProUGUI>();
-        dialog.text = dialogs[Random.Range(0, dialogs.Length)];
+        return dialogs[Random.Range(0, dialogs.Length)];
     }
-
-    public void GetDamage()
-    {
-        hp = hp - 1;
-    }
-
 }
