+++ Assets/Scenes/Villager1.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b4a162b719c76d64596761c7249e9e2b, type: 3}
+  m_Name: Villager1
+  m_EditorClassIdentifier: 
+  dialogs:
+  - AAAAA ayudaaaa
+  - me wa a moriiiirrrrr
+  - pipipipipip
+  timeMax: 6
