+++ Assets/Scenes/Villager3.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7f318b420dabb6e479327309f33bdec9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
