+++ Assets/Scenes/villager2.asset
@@ -10,11 +10,10 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: b4a162b719c76d64596761c7249e9e2b, type: 3}
-  m_Name: NPC Scriptable
+  m_Name: villager2
   m_EditorClassIdentifier: 
-  hp: 0
   dialogs:
-  - AAAAA ayudaaaa
-  - me wa a moriiiirrrrr
-  - pipipipipip
-  dialog: {fileID: 0}
+  - save meeeeeeeeeeee
+  - im gonna dieeeeeeeeee
+  - 'D:::::::::'
+  timeMax: 4
