+++ Assets/Scenes/villager2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 65a5eb0e922a6a84db95702ff07639b6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
